chore: standardize formatting and code-quality checks

+++ tests/unit/rabitqlib/utils/space_test.cpp
@@ -1,12 +1,15 @@
+#include "rabitqlib/utils/space.hpp"
+
 #include <gtest/gtest.h>
+
+#include <cmath>
+#include <vector>
+
+#include "rabitqlib/defines.hpp"
 #include "rabitqlib/simd/space_dispatch.hpp"
 #include "rabitqlib/utils/cpu_features.hpp"
-#include "rabitqlib/utils/space.hpp"
-#include "rabitqlib/defines.hpp"
-#include "test_helpers.hpp"
 #include "test_data.hpp"
-#include <vector>
-#include <cmath>
+#include "test_helpers.hpp"
 
 using namespace rabitqlib;
 using namespace rabitq_test;
@@ -62,7 +65,8 @@ TEST(ScalarQuantize, Uint8MatchesRoundedScalar) {
     std::vector<uint8_t> expected(dim);
 
     for (size_t i = 0; i < dim; ++i) {
-        float quantized = static_cast<float>((i * 7) % 251) + (static_cast<int>(i % 3) - 1) * 0.2F;
+        float quantized = static_cast<float>((i * 7) % 251) +
+                          static_cast<float>(static_cast<int>(i % 3) - 1) * 0.2F;
         input[i] = lo + delta * quantized;
         expected[i] = static_cast<uint8_t>(std::round((input[i] - lo) / delta));
     }
@@ -81,7 +85,8 @@ TEST(ScalarQuantize, Uint16MatchesRoundedScalar) {
     std::vector<uint16_t> expected(dim);
 
     for (size_t i = 0; i < dim; ++i) {
-        float quantized = static_cast<float>(1000 + i * 317) + (static_cast<int>(i % 5) - 2) * 0.1F;
+        float quantized = static_cast<float>(1000 + i * 317) +
+                          static_cast<float>(static_cast<int>(i % 5) - 2) * 0.1F;
         input[i] = lo + delta * quantized;
         expected[i] = static_cast<uint16_t>(std::round((input[i] - lo) / delta));
     }
@@ -95,8 +100,8 @@ TEST(ip16_fxu1_avx, ip_works) {
     srand(42);
     size_t dim = 64;
     float query[dim];
-    uint8_t codes[dim/8];
-    
+    uint8_t codes[dim / 8];
+
     for (size_t i = 0; i < dim; ++i) {
         query[i] = static_cast<float>(rand()) / RAND_MAX * 1000.0f;
     }
@@ -105,15 +110,17 @@ TEST(ip16_fxu1_avx, ip_works) {
         codes[i] = static_cast<uint8_t>(rand() % 256);
     }
 
-    ASSERT_NEAR(rabitqlib::excode_ipimpl::ip16_fxu1_avx(query, codes, dim), 15055.81f, 0.1f);
+    ASSERT_NEAR(
+        rabitqlib::excode_ipimpl::ip16_fxu1_avx(query, codes, dim), 15055.81f, 0.1f
+    );
 }
 
 TEST(ip64_fxu2_avx, ip_works) {
     srand(42);
-    size_t dim = 64*4;
+    size_t dim = 64 * 4;
     float query[dim];
-    uint8_t codes[dim/4];
-    
+    uint8_t codes[dim / 4];
+
     for (size_t i = 0; i < dim; ++i) {
         query[i] = static_cast<float>(rand()) / RAND_MAX * 1000.0f;
     }
@@ -121,7 +128,9 @@ TEST(ip64_fxu2_avx, ip_works) {
     for (size_t i = 0; i < dim / 4; ++i) {
         codes[i] = static_cast<uint8_t>(rand() % 256);
     }
-    ASSERT_NEAR(rabitqlib::excode_ipimpl::ip64_fxu2_avx(query, codes, dim), 217584.15f, 0.1f);
+    ASSERT_NEAR(
+        rabitqlib::excode_ipimpl::ip64_fxu2_avx(query, codes, dim), 217584.15f, 0.1f
+    );
 }
 
 TEST(ip_fxu8_avx, ip_works) {
